splitup renderer_base
