+++ src/actinia_core/rest/base/renderer_base.py
@@ -4,7 +4,7 @@
 # performance processing of geographical data that uses GRASS GIS for
 # computational tasks. For details, see https://actinia.mundialis.de/
 #
-# Copyright (c) 2016-2018 Sören Gebbert and mundialis GmbH & Co. KG
+# Copyright (c) 2016-2022 Sören Gebbert and mundialis GmbH & Co. KG
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -26,15 +26,12 @@ Render base classes
 """
 
 from flask_restful import reqparse
-from actinia_core.rest.ephemeral_processing import EphemeralProcessing
 from actinia_core.rest.base.resource_base import ResourceBase
-import os
 
 __license__ = "GPLv3"
 __author__ = "Sören Gebbert"
-__copyright__ = "Copyright 2016-2018, Sören Gebbert and mundialis GmbH & Co. KG"
-__maintainer__ = "Sören Gebbert"
-__email__ = "soerengebbert@googlemail.com"
+__copyright__ = "Copyright 2016-2022, Sören Gebbert and mundialis GmbH & Co. KG"
+__maintainer__ = "mundialis"
 
 
 REGION_PARAMETERS = {
@@ -182,52 +179,3 @@ class RendererBaseResource(ResourceBase):
         if "end_time" in args and args["end_time"] is not None:
             options["end_time"] = args["end_time"]
         return options
-
-
-class EphemeralRendererBase(EphemeralProcessing):
-
-    def __init__(self, *args):
-        EphemeralProcessing.__init__(self, *args)
-
-    def _setup_render_environment_and_region(self, options, result_file, legacy=True):
-        """Setup the render environment and create a g.region
-         process chain entry to setup the extent from the options.
-
-        Args:
-            options: The parser options that contain n, s, e and w entries for
-                     region settings
-            result_file: The resulting PNG file name
-            legacy: If True use the legacy process chain format for the
-                    g.region process definition
-
-        Returns:
-            A process chain entry of g.region
-
-        """
-        os.putenv("GRASS_RENDER_IMMEDIATE", "png")
-        os.putenv("GRASS_RENDER_WIDTH", str(options["width"]))
-        os.putenv("GRASS_RENDER_HEIGHT", str(options["height"]))
-        os.putenv("GRASS_RENDER_TRANSPARENT", "TRUE")
-        os.putenv("GRASS_RENDER_TRUECOLOR", "TRUE")
-        os.putenv("GRASS_RENDER_FILE", result_file)
-        os.putenv("GRASS_RENDER_FILE_READ", "TRUE")
-
-        if legacy is True:
-            pc = {"module": "g.region", "inputs": {}, "flags": "g"}
-            if options:
-                for key in options:
-                    if key in ["n", "s", "e", "w"]:
-                        value = options[key]
-                        pc["inputs"][key] = value
-        else:
-            pc = {"id": "2",
-                  "module": "g.region",
-                  "inputs": [],
-                  "flags": "g"}
-            if options:
-                for key in options:
-                    if key in ["n", "s", "e", "w"]:
-                        value = options[key]
-                        pc["inputs"].append({"param": key, "value": str(value)})
-
-        return pc
