+++ src/actinia_core/rest/strds_renderer.py
@@ -28,8 +28,9 @@ from flask import jsonify, make_response, Response
 from actinia_core.processing.actinia_processing.ephemeral.ephemeral_processing \
      import EphemeralProcessing
 from actinia_core.core.common.redis_interface import enqueue_job
-from actinia_core.rest.base.renderer_base import RendererBaseResource, \
-     EphemeralRendererBase
+from actinia_core.rest.base.renderer_base import RendererBaseResource
+from actinia_core.processing.actinia_processing.ephemeral.base.renderer_base \
+     import EphemeralRendererBase
 
 import os
 from flask_restful_swagger_2 import swagger
